Moved the modules folder out as it behaves more like and API so belongs seperate from business logic
